project/core/custom_storages.py
# from fnmatch import fnmatch
from django.conf import settings
from storages.backends.s3boto3 import S3Boto3Storage

class StaticStorage(S3Boto3Storage):
    access_key = settings.AWS_S3_ACCESS_KEY_ID
    secret_key = settings.AWS_SECRET_ACCESS_KEY
    bucket_name = settings.AWS_STORAGE_BUCKET_NAME
    region_name = settings.AWS_S3_BUCKET_REGION
    location = 'static'

class MediaStorage(S3Boto3Storage):
    access_key = settings.AWS_S3_ACCESS_KEY_ID
    secret_key = settings.AWS_SECRET_ACCESS_KEY
    bucket_name = settings.AWS_STORAGE_BUCKET_NAME
    region_name = settings.AWS_S3_BUCKET_REGION
    location = 'media'