project/requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-browser-reload==1.12.1
django-cotton==0.9.32
django-extensions==3.2.3
django-filter==24.2
django-htmx==1.17.3
django-storages==1.14.3
django-unfold==0.33.0
djangorestframework==3.15.1
Faker==24.14.0
fire==0.6.0
gunicorn==22.0.0
idna==3.7
kombu==5.3.7
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
rav==0.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
termcolor==2.4.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
